item and location IDs are being taken as null; the error=> org.postgresql.util.PSQLException: ERROR: null value in column "item_id" violates not-null constraint   Detail: Failing row contains (5, null, null, 10, INTRANSIT).

+++ src/main/java/com/mdsujan/restPostgres/response/SupplyResponse.java
@@ -16,7 +16,7 @@ public class SupplyResponse {
 
     AllowedSupplyTypes type;
 
-    Integer qty;
+    Long qty;
 
     Long itemId;
